laboratories/laboratory_3/1019147265/api_gateway/requirements.txt
Flask
requests
PyJWT
